lakshmi-api/requirements.txt
click==7.1.2
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.68.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
motor==3.3.2
passlib==1.7.4
pathspec==0.8.1
pyasn1==0.4.8
pydantic==1.9.1
pymongo==4.6.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
regex==2022.8.17
rsa==4.9
six==1.16.0
toml==0.10.2
uvicorn==0.12.2
uvloop==0.16.0